Added forceOverwrite option

+++ server/src/main/java/pro/prysm/orion/server/Native.java
@@ -12,22 +12,33 @@ public class Native {
 
     /**
      * Loads a native library from the resources of this Jar
-     * @param name Library to load (without extension)
+     * @param name Library to load
      * @throws ResourceNotFoundException Library not found in resources
      * @throws IOException Cannot move library out of Jar
      */
-    public static void load(String name) throws ResourceNotFoundException, IOException {
+    public static void load(String name) throws IOException, ResourceNotFoundException {
+        load(name, false);
+    }
+
+    /**
+     * Loads a native library from the resources of this Jar
+     * @param name Library to load
+     * @param forceOverwrite Force overwrite the library file
+     * @throws ResourceNotFoundException Library not found in resources
+     * @throws IOException Cannot move library out of Jar
+     */
+    public static void load(String name, boolean forceOverwrite) throws ResourceNotFoundException, IOException {
         String os = System.getProperty("os.name");
         if (os.contains("Linux")) name = name + ".so";
         else if (os.contains("Mac OS")) name = name + ".dylib";
         File libDir = new File("lib");
         if (!libDir.exists()) libDir.mkdirs();
         File object = new File("lib", name);
-        // if (!object.exists()) {
+        if (forceOverwrite || !object.exists()) {
             InputStream is = Orion.class.getClassLoader().getResourceAsStream(name);
             if (is == null) throw new ResourceNotFoundException("Could not find lib " + name + " in jar");
             Files.copy(is, object.toPath(), StandardCopyOption.REPLACE_EXISTING);
-        // }
+        }
         System.load(object.getAbsolutePath());
     }
 }
